// app/data/enum.ts
// SKILLS
export enum Proficiency {
    BEGINNER = "Beginner",
    FUNCTIONAL = "Functional",
    INTERMEDIATE = "Intermediate",
    ADVANCE = "Advance",
    PROFICIENT = "Proficient"
}

export enum Preference {
    CORE_SKILL = "Core Skills",
    SUB_SKILL = "Sub Skills",
    DESIGN = "Design Skills",
    HARDWARE = "Hardware",
    SOFT_SKILLS = "Soft Skills",
    OTHER = "Other"
}

export enum Category {
    FUNDAMENTAL = "Fundamentals",
    FRAMEWORK = "Frameworks",
    CLI = "CLI",
    LANGUAGE = "Language",
    BACKEND = "Backend",
    DATABASE = "Database",
    ORM = "ORMs",
    STACK = "Stacks",
    TOOLS = "Tools",
    VERSION_CONTROL = "CI/CD & Deployment",
    GRAPHICS = "Graphics",
    SOCIAL_MEDIA = "Social Media",
    OS = "Operating System",
    HARDWARE = "Hardware",
    OTHER = "Other"
}